refactor(frontend): moved custom files from /ui
