+++ new/src/frontend/acc2/src/components/home/sections/compute.tsx
@@ -10,7 +10,7 @@ import z from "zod";
 import { useFileUploadMutation } from "@acc2/hooks/mutations/files";
 import { toast } from "sonner";
 import { useComputationMutation } from "@acc2/hooks/mutations/use-computation-mutation";
-import { Busy } from "@acc2/components/ui/busy";
+import { Busy } from "@acc2/components/shared/busy";
 import { handleApiError } from "@acc2/api/base";
 import { useSetupMutation } from "@acc2/hooks/mutations/use-setup-mutation";
